Backend/requirements.txt
fastapi==0.109.2
mangum==0.17.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
email-validator==2.1.0.post1
httpx==0.27.0
python-jose[cryptography]==3.3.0


boto3==1.34.49
aws-xray-sdk==2.14.0

python-dateutil==2.8.2
tenacity==8.2.3
cachetools==5.3.2
requests==2.32.3
aiohttp==3.10.11
groq

deep-translator==1.11.4

numpy==2.1.3
Pillow==10.2.0

apscheduler==3.10.4
holidays==0.57
beautifulsoup4==4.12.3

bcrypt==4.1.2
passlib[bcrypt]==1.7.4
